backend/requirements.txt
Django==1.11.29
djangorestframework==3.14.0
django-cors-headers
PyJWT==2.8.0
django-filter